Feature/yam 14 create create train screen

+++ app/src/main/java/com/sedilant/yambol/data/entities/TrainTaskEntity.kt
@@ -8,7 +8,7 @@ import androidx.room.PrimaryKey
 )
 data class TrainTaskEntity(
     @PrimaryKey(autoGenerate = true)
-    val trainingTaskId: Int = 0,
+    val trainingTaskId: Long = 0,
     val name: String,
     val numberOfPlayers: Int,
     val concept: String,
