+++ app/src/main/java/com/sedilant/yambol/domain/models/TrainWithTaskDomainModel.kt
@@ -3,7 +3,7 @@ package com.sedilant.yambol.domain.models
 import java.util.Date
 
 data class TrainWithTaskDomainModel(
-    val trainId: Int,
+    val trainId: Long,
     val date: Date,
     val time: Float,
     val concepts: List<String>,
